Restructure directories and modularize
