+++ cerebro/modelo/avaliador_modelo.py
