+++ cerebro/modelo/treino_modelo.py
